requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
